convert FrameQueue elements' double frm_pts to int64_t frm_pts_usec

+++ libavPlayer/VideoState.h
@@ -45,6 +45,6 @@
     // Video
 
     int width, height;
-    double last_shown_video_frame_pts;
+    int64_t last_shown_video_frame_pts;
 }
 @end;
